add CI runs for macos/nix and move tests into script
Adds multiple new test runs, which will be filled out by the upcoming commits.
This also refactors the runs into segments so that they can be composed to
minimize repetition where feasible.

The runs include:
- in-place linux build (the current CI run)
- an in-place macOS build (this build is broken--it's just a placeholder
  since you indicated interest in a non-nix macOS build)
- two pairs of macos/linux Nix builds (one w/ nix-shell, one w/ nix-build)

Only the in-place linux build is required; none of the others will currently
block build.